refactor : 카테고리 등록 / 수정 시 중복 체크 fix : 중복 엔터티 선언 수정
