+++ backend/src/main/java/com/pigma/harusari/category/command/domain/repository/CategoryCommandRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryCommandRepository extends JpaRepository<Category, Long> {
 
     // 카테고리 이름 중복 확인
     boolean existsByMemberIdAndCategoryName(Long memberId, String categoryName);
