+++ backend/src/main/java/com/pigma/harusari/category/command/domain/service/CategoryCommandServiceImpl.java
@@ -3,7 +3,7 @@ package com.pigma.harusari.category.command.domain.service;
 import com.pigma.harusari.category.command.application.dto.request.CategoryCreateRequest;
 import com.pigma.harusari.category.command.application.dto.request.CategoryUpdateRequest;
 import com.pigma.harusari.category.command.domain.aggregate.Category;
-import com.pigma.harusari.category.command.domain.repository.CategoryRepository;
+import com.pigma.harusari.category.command.domain.repository.CategoryCommandRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,11 +12,15 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class CategoryCommandServiceImpl implements CategoryCommandService {
 
-    private final CategoryRepository categoryRepository;
+    private final CategoryCommandRepository categoryCommandRepository;
 
     @Override
     @Transactional
     public Long createCategory(CategoryCreateRequest request) {
+        if (categoryCommandRepository.existsByMemberIdAndCategoryName(request.getMemberId(), request.getCategoryName())) {
+            throw new IllegalStateException("이미 동일한 이름의 카테고리가 존재합니다.");
+        }
+
         Category category = Category.builder()
                 .memberId(request.getMemberId())
                 .categoryName(request.getCategoryName())
@@ -24,14 +28,20 @@ public class CategoryCommandServiceImpl implements CategoryCommandService {
                 .completed(false)
                 .build();
 
-        return categoryRepository.save(category).getCategoryId();
+        return categoryCommandRepository.save(category).getCategoryId();
     }
 
     @Override
     @Transactional
     public void updateCategory(Long categoryId, Long memberId, CategoryUpdateRequest request) {
-        Category category = categoryRepository.findByCategoryIdAndMemberId(categoryId, memberId)
+        Category category = categoryCommandRepository.findByCategoryIdAndMemberId(categoryId, memberId)
                 .orElseThrow(() -> new IllegalArgumentException("카테고리를 찾을 수 없습니다."));
+
+        // 같은 이름이지만 본인 카테고리가 아닌 경우만 체크
+        if (!category.getCategoryName().equals(request.getCategoryName()) &&
+                categoryCommandRepository.existsByMemberIdAndCategoryName(memberId, request.getCategoryName())) {
+            throw new IllegalStateException("이미 동일한 이름의 카테고리가 존재합니다.");
+        }
 
         category.updateCategoryDetails(request.getCategoryName(), request.getColor());
     }
@@ -39,7 +49,7 @@ public class CategoryCommandServiceImpl implements CategoryCommandService {
     @Override
     @Transactional
     public void completeCategory(Long categoryId, Long memberId) {
-        Category category = categoryRepository.findByCategoryIdAndMemberId(categoryId, memberId)
+        Category category = categoryCommandRepository.findByCategoryIdAndMemberId(categoryId, memberId)
                 .orElseThrow(() -> new IllegalArgumentException("카테고리를 찾을 수 없습니다."));
 
         category.completeCategory();
@@ -52,10 +62,10 @@ public class CategoryCommandServiceImpl implements CategoryCommandService {
             throw new IllegalArgumentException("삭제 확인 문구가 일치하지 않습니다.");
         }
 
-        Category category = categoryRepository.findByCategoryIdAndMemberId(categoryId, memberId)
+        Category category = categoryCommandRepository.findByCategoryIdAndMemberId(categoryId, memberId)
                 .orElseThrow(() -> new IllegalArgumentException("카테고리를 찾을 수 없습니다."));
 
         // 연결된 일정 삭제는 추후 처리
-        categoryRepository.delete(category);
+        categoryCommandRepository.delete(category);
     }
 }
