--- backend/src/main/java/com/pigma/harusari/category/command/repository/CategoryCommandRepository.java
@@ -1,7 +1,0 @@
-package com.pigma.harusari.category.command.repository;
-
-import com.pigma.harusari.category.command.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryCommandRepository extends JpaRepository<Category, Long> {
-}
