+++ backend/src/main/java/com/pigma/harusari/category/query/service/CategoryQueryServiceImpl.java
@@ -1,8 +1,7 @@
-package com.pigma.harusari.category.query.service.category.query.service;
+package com.pigma.harusari.category.query.service;
 
 import com.pigma.harusari.category.query.dto.response.CategoryResponse;
 import com.pigma.harusari.category.query.mapper.CategoryMapper;
-import com.pigma.harusari.category.query.service.category.query.service.CategoryQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
