+++ backend/src/main/java/com/pigma/harusari/user/command/service/UserCommandServiceImpl.java
@@ -1,8 +1,9 @@
 package com.pigma.harusari.user.command.service;
 
-import com.pigma.harusari.category.command.dto.CategoryCreateRequest;
-import com.pigma.harusari.category.command.entity.Category;
-import com.pigma.harusari.category.command.repository.CategoryCommandRepository;
+
+import com.pigma.harusari.category.command.application.dto.request.CategoryCreateRequest;
+import com.pigma.harusari.category.command.domain.aggregate.Category;
+import com.pigma.harusari.category.command.domain.repository.CategoryCommandRepository;
 import com.pigma.harusari.user.command.dto.SignUpRequest;
 import com.pigma.harusari.user.command.entity.Gender;
 import com.pigma.harusari.user.command.entity.Member;
